Shapoorji Pallonji Worli Mumbai
Shapoorji Pallonji Worli has been designed to meet the living requirements of a diverse population with comfortable homes and apartments that offer both quality and esthetics, while also providing easy access to facilities.

Living in a city like Mumbai, where you barely have time to hang out with friends and family, is tough. This is why Shapoorji Pallonji Worli Mumbai offers modern apartments in the heart of Mumbai so that you can enjoy a sense of peacefulness, yet be close to everything. All the residences are 1, 2 & 3 BHK Luxury Residences with premium facilities and utilities.

For More Details :
Visit - https://www.shapoorjipallonji.co.in/upcoming/mumbai/shapoorji-pallonji-worli/